feat: 停留所候補リストをバックエンドAPIから取得しビルド時にキャッシュ
- src/lib/busStops.ts: API取得関数を作成
  - https://prometheus-h24i.onrender.com/combus/stops から取得
  - lon → lng に変換、エラー時はフォールバック
- src/components/MapView.tsx: Client Componentに分離
  - 旧 page.tsx のロジックを移動
- src/app/page.tsx: Server Componentに変更
  - ビルド時にAPI取得、静的レンダリング
- CLAUDE.md: プロジェクト知識ベースを追加
- .claudeignore: 不要ファイルを除外してコンテキスト最適化
